+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 **This version is the `Quality of Life` improvement version.**
 
+- [fuji] feature: make the `/fuji` command a shortcut to `/fuji inspect modules`.
 - [fuji] feature: display the `relative path` for `/fuji inspect configurations`. (It's shorter, if not harm the readability)
 - [core] feature: enhance the `search function` in `all paged GUI`, now we will search any text you can see. (Much better to search a `keyword` inside Item)